requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.8
redis>=5.0.1
groq>=0.4.0
sqlalchemy>=2.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
alembic>=1.12.0 